+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 2.31 branch
 ### 2.31.1
 * fixed little bug for `ControllerTestCase::assert*Authorized()` methods;
+* little fixes for controllers;
 * updated for php-tools 1.7.1.
 
 ### 2.31.0